fixes for python3

+++ lib/lib.py
@@ -9,6 +9,6 @@ def fft_convolve2d(x,y):
     fr2 = fft2(np.flipud(np.fliplr(y)))
     m,n = fr.shape
     cc = np.real(ifft2(fr*fr2))
-    cc = np.roll(cc, -m/2+1,axis=0)
-    cc = np.roll(cc, -n/2+1,axis=1)
+    cc = np.roll(cc, - int(m / 2) + 1, axis=0)
+    cc = np.roll(cc, - int(n / 2) + 1, axis=1)
     return cc
